requirements.txt
pandas==1.3.4
psycopg2-binary==2.9.9
python-binance==1.0.22
python-dotenv==1.0.1
schedule==0.6.0
SQLAlchemy==1.4.35
